refactor: DetailPage/ 스타일드 컴포넌트 분리
