+++ fe/src/Component/DetailPage/DetailPageDesc/Explaination/Explaination.jsx
@@ -1,13 +1,6 @@
-import styled from "styled-components";
+import ExplainationDiv from "./Explanation.styled";
 import Images from "./Images/Images";
 import Order from "./Order/Order";
-
-const ExplainationDiv = styled.div`
-  display: flex;
-  padding: 0 48px;
-  height: 472px;
-  justify-content: space-between;
-`;
 
 const Explaination = () => {
   return (
