+++ fe/src/Component/DetailPage/DetailPageDesc/Explaination/Explanation.styled.js
@@ -1,0 +1,10 @@
+import styled from "styled-components";
+
+const ExplainationDiv = styled.div`
+  display: flex;
+  padding: 0 48px;
+  height: 472px;
+  justify-content: space-between;
+`;
+
+export default ExplainationDiv;
